requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
python-socketio==5.11.0
python-engineio==4.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
geopy==2.4.1
pandas==2.1.4
python-dateutil==2.8.2
eventlet==0.33.3